Navigation buttons now bring the user to a new screen - profile
