+++ components/ProfileScreen.js
@@ -1,0 +1,8 @@
+import * as React from 'react';
+
+export default function ProfileScreen({ navigation }) {
+  return (
+      <p>Profile Screen</p>
+
+  );
+}
